The filling of the mold
Then on 21th October 2015 we sterilised the mold with bleach and then
alcohol. After that we filled it “painted” it with a blend of malt agar.
Then we filled the mold with straw that has been sterilised in the
pressure cooker for 20 minutes. The straw will provide for structure (so
that the resulting head is strong enough to show on our final night) and
food for the mycelium.
We then added the bits of mycelium which I cultivated before. The big
question will be if the mycelium will grow fast enough to fill up the
mold before the 17th when our presentations are planned.